Actually creating the context providers for config and alt
